+++ tempodb/backend/cache/memcached/memcached.go
@@ -1,14 +1,12 @@
 package memcached
 
 import (
-	"context"
 	"time"
 
 	"github.com/go-kit/kit/log"
 	"github.com/prometheus/client_golang/prometheus"
 
 	cortex_cache "github.com/cortexproject/cortex/pkg/chunk/cache"
-	"github.com/grafana/tempo/tempodb/backend/cache"
 )
 
 type Config struct {
@@ -17,11 +15,7 @@ type Config struct {
 	TTL time.Duration `yaml:"ttl"`
 }
 
-type Client struct {
-	client *cortex_cache.Memcached
-}
-
-func NewClient(cfg *Config, logger log.Logger) cache.Client {
+func NewClient(cfg *Config, cfgBackground *cortex_cache.BackgroundConfig, logger log.Logger) cortex_cache.Cache {
 	if cfg.ClientConfig.MaxIdleConns == 0 {
 		cfg.ClientConfig.MaxIdleConns = 16
 	}
@@ -38,26 +32,7 @@ func NewClient(cfg *Config, logger log.Logger) cache.Client {
 		BatchSize:   0, // we are currently only requesting one key at a time, which is bad.  we could restructure Find() to batch request all blooms at once
 		Parallelism: 0,
 	}
-	return &Client{
-		client: cortex_cache.NewMemcached(memcachedCfg, client, "tempo", prometheus.DefaultRegisterer, logger),
-	}
-}
+	cache := cortex_cache.NewMemcached(memcachedCfg, client, "tempo", prometheus.DefaultRegisterer, logger)
 
-// Store implements cache.Store
-func (m *Client) Store(ctx context.Context, key string, val []byte) {
-	m.client.Store(ctx, []string{key}, [][]byte{val})
-}
-
-// Fetch implements cache.Fetch
-func (m *Client) Fetch(ctx context.Context, key string) []byte {
-	found, vals, _ := m.client.Fetch(ctx, []string{key})
-	if len(found) > 0 {
-		return vals[0]
-	}
-	return nil
-}
-
-// Shutdown implements cache.Shutdown
-func (m *Client) Shutdown() {
-	m.client.Stop()
+	return cortex_cache.NewBackground("tempo", *cfgBackground, cache, prometheus.DefaultRegisterer)
 }